+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/Diffusion/Node/MiniProtocols.hs
@@ -432,7 +432,6 @@ applications debugTracer nodeKernel
       -> do labelThisThread "BlockFetchClient"
             bracketFetchClient (nkFetchClientRegistry nodeKernel)
                                UnversionedProtocol
-                               (const NotReceivingTentativeBlocks)
                                remoteAddress
                                $ \clientCtx ->
               runPeerWithLimits